+++ app/package.json
@@ -20,6 +20,8 @@
     "@sentry/react": "^6.6.0",
     "@sentry/tracing": "^6.6.0",
     "browserslist-useragent-regexp": "^3.0.0",
+    "buffer": "^6.0.3",
+    "canvas-webworker": "^0.0.3",
     "date-fns": "^2.21.3",
     "deep-freeze": "~0.0.1",
     "fend-wasm-web": "~0.1.26",
@@ -31,6 +33,7 @@
     "mousetrap": "^1.6.1",
     "opus-recorder": "^4.1.0",
     "paste-image": "^0.0.3",
+    "path-browserify": "^1.0.1",
     "pdfobject": "^2.1.1",
     "re-resizable": "^6.9.0",
     "react": "^17.0.2",
@@ -49,6 +52,7 @@
     "save-as": "^0.1.8",
     "semver": "^7.3.4",
     "showdown": "^1.8.6",
+    "stream-browserify": "^3.0.0",
     "timers-browserify": "^2.0.12",
     "typescript": "~4.4.3",
     "upad-parse": "^7.3.0",
